requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.5.0
catboost==1.2.7
mlxtend==0.23.1
matplotlib==3.9.2
mlflow==2.16.2
shap==0.46.0
lime==0.2.0.1
transformers==4.45.2
sentencepiece==0.2.0
accelerate==1.0.1
evaluate==0.4.3
sacrebleu==2.4.3
sentence-transformers==3.1.1
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
tqdm==4.66.5
httpx==0.27.2
plotly==5.24.1
ipykernel==6.29.5
jupyter==1.1.1
pre-commit==4.0.1
black==24.8.0
isort==5.13.2
flake8==7.1.1
